3/27 11:17 chenxuan

+++ STM32Smart_home_health/Smart_home_health_Keil/BSP/LM2904/bsp_lm2904_driver.h
@@ -5,7 +5,7 @@
 
 void lm2904_init(void);
 
-void lm2904_start(uint16_t* data,uint16_t len);
+void lm2904_start(uint32_t* data,uint32_t len);
 void lm2904_stop(void);
 float adc_to_db(uint32_t adc_value) ;
